Add Modules rule category
