+++ src/rules/best_practices/mod.rs
@@ -1,2 +1,1 @@
 pub mod filesystem;
-pub mod modules_and_programs;
