+++ src/rules/modules/external_functions.rs
@@ -1,11 +1,8 @@
 use crate::{Method, Rule, Violation};
 use tree_sitter::{Node, Query};
 
-/// Defines rules that check whether functions and subroutines are defined within modules,
-/// submodules, or interfaces. Also ensures that `use` statements are used correctly.
-
-// Define functions and subroutines in modules
-// -------------------------------------------
+/// Defines rules that check whether functions and subroutines are defined within modules (or one
+/// of a few acceptable alternatives).
 
 fn external_fucntion(root: &Node, src: &str) -> Vec<Violation> {
     let mut violations = Vec::new();
@@ -37,59 +34,6 @@ impl Rule for ExternalFunction {
         Functions and subroutines should be contained within (sub)modules or programs.
         Fortran compilers are unable to perform type checks and conversions on functions
         defined outside of these scopes, and this is a common source of bugs.
-        "
-    }
-}
-
-// Always follow 'use' with 'only'
-// -------------------------------
-
-fn use_all(root: &Node, src: &str) -> Vec<Violation> {
-    let mut violations = Vec::new();
-    // Search for 'use' clause, and optionally an 'only' clause, capturing both.
-    let query_txt = "(use_statement (included_items)? @only) @use";
-    let query = Query::new(&tree_sitter_fortran::language(), query_txt).unwrap();
-    let only_index = query.capture_index_for_name("only").unwrap();
-    let mut cursor = tree_sitter::QueryCursor::new();
-    for captures in cursor
-        .matches(&query, *root, src.as_bytes())
-        .map(|x| x.captures)
-    {
-        // If the @only capture isn't found, record a violation.
-        if !captures.iter().any(|&x| x.index == only_index) {
-            // The only remaining capture must be the use clause
-            for capture in captures {
-                let msg = "'use' statement missing 'only' clause";
-                violations.push(Violation::from_node(msg, &capture.node));
-            }
-        }
-    }
-    violations
-}
-
-pub struct UseAll {}
-
-impl Rule for UseAll {
-    fn method(&self) -> Method {
-        Method::Tree(use_all)
-    }
-
-    fn explain(&self) -> &str {
-        "
-        When using a module, it is recommended to add an 'only' clause to specify which
-        components you intend to use:
-
-        ```
-        ! Not recommended
-        use, intrinsic :: iso_fortran_env
-
-        ! Better
-        use, intrinsic :: iso_fortran_env, only: int32, real64
-        ```
-
-        This makes it easier for programmers to understand where the symbols in your
-        code have come from, and avoids introducing many unneeded components to your
-        local scope.
         "
     }
 }
@@ -144,19 +88,5 @@ mod tests {
             end module
             ";
         test_tree_method(external_fucntion, source, None);
-    }
-
-    #[test]
-    fn test_use_all() {
-        let source = dedent(
-            "
-            module my_module
-                use, intrinsic :: iso_fortran_env, only: real32
-                use, intrinsic :: iso_c_binding
-            end module
-            ",
-        );
-        let violation = violation!("'use' statement missing 'only' clause", 4, 5);
-        test_tree_method(use_all, source, Some(vec![violation]));
     }
 }
