+++ src/rules/modules/mod.rs
@@ -1,0 +1,6 @@
+pub mod external_functions;
+pub mod use_statements;
+
+// TODO should be private by default, with explicit public interface
+// TODO prefer 'end module {name}'
+// TODO function is not used within a module and is not public
